Add UI & game scaling with window
Added:
- SceneManager::SetShowAllBounds()
- Scene::GetElements()

Updated:
- All debug keyboard shortcuts are now useable only in debug mode
- GUI and game is now scaling with window
- ToggleShowFocused() now switches between none/focused/all
- Settings default window style is now Titlebar | Close

Fixed:
- Minor build warnings
- Some debug options weren't working if set before entering game

+++ RoguelikeGame.Main/Engine/Helpers/MathHelper.cpp
@@ -8,7 +8,7 @@ float MathHelper::GetAngleBetweenPoints(const sf::Vector2f& first, const sf::Vec
 
 float MathHelper::GetDistanceBetweenPoints(const sf::Vector2f& first, const sf::Vector2f& second)
 {
-    return sqrt(pow(first.x - second.x, 2) + pow(first.y - second.y, 2));
+    return sqrtf(powf(first.x - second.x, 2) + powf(first.y - second.y, 2));
 }
 
 sf::Vector2f MathHelper::GetPointFromAngle(const sf::Vector2f& start, float angle, float radius)
